feat: 각 도메인의 repository 메서드 구현 (쿼리)

+++ mjuBank/src/main/java/com/example/mjuBank/repository/CustomerRepository.java
@@ -3,8 +3,7 @@ package com.example.mjuBank.repository;
 
 import com.example.mjuBank.domain.Customer;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
 }
